+++ src/engine/options.rs
@@ -1,5 +1,6 @@
 pub mod defaults {
     pub const HASH_SIZE: usize = 256;
+    pub const THREADS: usize = 1;
     pub const ENABLE_LOGGING: bool = false;
     pub const MOVE_OVERHEAD: usize = 0;
 }
@@ -7,6 +8,7 @@ pub mod defaults {
 #[derive(Debug, Clone)]
 pub struct EngineOptions {
     pub hash_size: usize,
+    pub threads: usize,
     pub enable_logging: bool,
 
     // Account for the possibility that there's some overhead making the move
@@ -18,6 +20,7 @@ impl Default for EngineOptions {
     fn default() -> Self {
         Self {
             hash_size: defaults::HASH_SIZE,
+            threads: defaults::THREADS,
             enable_logging: defaults::ENABLE_LOGGING,
             move_overhead: defaults::MOVE_OVERHEAD,
         }